LPS-163431 Move pagination attributes handling to base display class, simplify
